requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0

# Additional utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Optional dependencies for enhanced functionality
aiofiles==23.2.1
jinja2==3.1.2